Preprocessor seperation from __init__ and added directive custom hook
